+++ Joker/src/components/modals/loading.tsx
@@ -5,28 +5,27 @@ import {Button, StatusBar, StyleSheet, View} from 'react-native';
 import {Waiting} from 'src/components/lottie';
 import {Text} from 'src/components/ui';
 import {hideModal} from 'src/helpers';
-import {
-  IS_ANDROID,
-  LIGHT_BG_1,
-  LIGHT_GRAPHIC_GREEN_2,
-  LIGHT_TEXT_BASE_3,
-} from 'src/variables';
+import {useTheme, useThemeObject} from 'src/hooks';
+import {ColorTheme} from 'src/types';
+import {IS_ANDROID} from 'src/variables';
 
 export type LoadingModalProps = {
   text?: string;
 };
 export function LoadingModal({text}: LoadingModalProps) {
+  const {color} = useTheme().theme;
+  const styles = useThemeObject(createStyles);
   useEffect(() => {
     if (IS_ANDROID) {
-      StatusBar.setBackgroundColor(LIGHT_GRAPHIC_GREEN_2);
-      return () => StatusBar.setBackgroundColor(LIGHT_BG_1);
+      StatusBar.setBackgroundColor(color.graphicGreen2);
+      return () => StatusBar.setBackgroundColor(color.bg1);
     }
   }, []);
   return (
-    <View style={page.container}>
-      <Waiting style={page.waiting} />
+    <View style={styles.container}>
+      <Waiting style={styles.waiting} />
       {text && (
-        <Text t4 style={page.text}>
+        <Text t4 style={styles.text}>
           {text}
         </Text>
       )}
@@ -34,14 +33,16 @@ export function LoadingModal({text}: LoadingModalProps) {
     </View>
   );
 }
-
-const page = StyleSheet.create({
-  container: {
-    justifyContent: 'center',
-    alignItems: 'center',
-    flex: 1,
-    backgroundColor: LIGHT_GRAPHIC_GREEN_2,
-  },
-  text: {color: LIGHT_TEXT_BASE_3, width: 230, textAlign: 'center'},
-  waiting: {marginBottom: 40},
-});
+const createStyles = (color: ColorTheme) => {
+  const styles = StyleSheet.create({
+    container: {
+      justifyContent: 'center',
+      alignItems: 'center',
+      flex: 1,
+      backgroundColor: color.graphicGreen2,
+    },
+    text: {color: color.textBase3, width: 230, textAlign: 'center'},
+    waiting: {marginBottom: 40},
+  });
+  return styles;
+};
