+++ Joker/src/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './use-typed-navigation';
 export * from './use-typed-route';
 export * from './use-theme-object';
+export * from './use-theme';
